refactor: update how set-password page is displayed
